Reorganized Again + Implemented Camera using Cinemachine
Fixed organization of Scripts folder such that Unity recognizes its location. Additionally, reworked Camera control + movement to use Cinemachine. This is a package that comes with Unity that allows us to have a production level camera control. Reworked character movement such that character will now face the direction it last travelled in while allowing movement to be relative to the camera's direction.