Added Basic animation blending with movement + PS4 Controller Input
Added a temporary animation controller that blends the idle, walk, and run animations for the character and tied the rate of blend to the character's movement. Additionally, I have updated the Input Manager to allow for PS4 controller input. Should work with PS4 controllers connected with bluetooth or wireless through a program called DS4Windows. Not sure if it works without DS4Windows running