Added Current Model + Rig
Added current version of Model & Rig. Will likely be updated later. Note that if the bones in the armature get editted or renamed, animations WILL break, so save your keyframes as a separate, concept art style like design on a separate file just in case we need to reimplement them.